+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/mod.rs
@@ -1,8 +1,9 @@
 use cdk_framework::act::CandidTypes;
+use swc_ecma_ast::TsTypeAliasDecl;
 
-use crate::ts_ast::TsAst;
+use crate::ts_ast::{source_map::SourceMapped, TsAst};
 
-mod ts_ast;
+mod services;
 mod ts_type_alias_decl;
 
 pub mod azle_type_ref;
@@ -24,5 +25,15 @@ impl TsAst {
             type_aliases,
             variants,
         }
+    }
+
+    pub fn extract_candid_types<F, T>(&self, extractor: F) -> Vec<T>
+    where
+        F: Fn(&SourceMapped<TsTypeAliasDecl>) -> Option<T>,
+    {
+        self.ts_type_alias_decls()
+            .iter()
+            .filter_map(|ts_type_alias_decl| extractor(ts_type_alias_decl))
+            .collect()
     }
 }
