+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/module.rs
@@ -1,11 +1,9 @@
 use swc_common::SourceMap;
-use swc_ecma_ast::{
-    ClassDecl, Decl, ExportDecl, Expr, FnDecl, Module, ModuleDecl, ModuleItem, Stmt,
-};
+use swc_ecma_ast::{Decl, ExportDecl, FnDecl, Module, ModuleDecl, ModuleItem, Stmt};
 
 use crate::{
     canister_method::{errors, module_item::ModuleItemHelperMethods, Annotation},
-    ts_ast::{source_map::SourceMapped, AzleFnDecl, AzleTypeAliasDecl, GetName},
+    ts_ast::{source_map::SourceMapped, AzleFnDecl, AzleTypeAliasDecl},
 };
 
 pub trait ModuleHelperMethods {
@@ -14,10 +12,6 @@ pub trait ModuleHelperMethods {
     fn get_export_decls(&self) -> Vec<ExportDecl>;
     fn get_azle_type_alias_decls<'a>(&'a self, source_map: &'a SourceMap)
         -> Vec<AzleTypeAliasDecl>;
-    fn get_service_class_declarations<'a>(
-        &'a self,
-        source_map: &'a SourceMap,
-    ) -> Vec<SourceMapped<ClassDecl>>;
 }
 
 impl ModuleHelperMethods for Module {
@@ -163,42 +157,6 @@ impl ModuleHelperMethods for Module {
             .collect();
 
         vec![stmt_azle_type_alias_decls, export_azle_type_alias_decls].concat()
-    }
-
-    fn get_service_class_declarations<'a>(
-        &'a self,
-        source_map: &'a SourceMap,
-    ) -> Vec<SourceMapped<ClassDecl>> {
-        self.body.iter().fold(vec![], |mut acc, module_item| {
-            // acc is mut because SourceMapped<FnDecl> can't be cloned, which is
-            // necessary to do something like:
-            // return vec![acc, vec![SourceMapped::new(class_decl, source_map)]].concat();
-
-            let decl_opt = match module_item {
-                ModuleItem::ModuleDecl(decl) => match decl {
-                    ModuleDecl::ExportDecl(export_decl) => Some(&export_decl.decl),
-                    _ => None,
-                },
-                ModuleItem::Stmt(stmt) => match stmt {
-                    Stmt::Decl(decl) => Some(decl),
-                    _ => None,
-                },
-            };
-
-            if let Some(decl) = decl_opt {
-                if let Decl::Class(class_decl) = decl {
-                    if let Some(super_class) = &class_decl.class.super_class {
-                        if let Expr::Ident(ident) = &**super_class {
-                            if ident.get_name() == "Service" {
-                                acc.push(SourceMapped::new(class_decl, source_map))
-                            }
-                        }
-                    }
-                }
-            }
-
-            acc
-        })
     }
 }
